Make clonefile optional
clonefile fails to build on Netlify:

    clonefile.c:47:58: error: ‘FICLONE’ undeclared

… which is unexpected, but because this is a regression, Nanoc can now work without clonefile as well (and only making use of it when it’s available).

+++ nanoc-core/lib/nanoc/core/item_rep_writer.rb
@@ -79,11 +79,13 @@ module Nanoc
 
       def smart_cp(from, to)
         # Try clonefile
-        FileUtils.rm_f(to)
-        begin
-          res = Clonefile.always(from, to)
-          return if res
-        rescue Clonefile::UnsupportedPlatform, Errno::ENOTSUP, Errno::EXDEV, Errno::EINVAL
+        if defined?(Clonefile)
+          FileUtils.rm_f(to)
+          begin
+            res = Clonefile.always(from, to)
+            return if res
+          rescue Clonefile::UnsupportedPlatform, Errno::ENOTSUP, Errno::EXDEV, Errno::EINVAL
+          end
         end
 
         # Try with hardlink
